refactor(reference): migrate measurement to attestation::reference
Move launch digest, ID block, and SEV hash reference calculation into
attestation::reference and remove the legacy measurement and vmsa
modules.
